Creating instrument type thus creating the skeleton for other entities [skip ci]
