+++ cms/src/web/profiles/open_pension/modules/open_pension_reclamation/src/Entity/InstrumentType.php
@@ -34,7 +34,7 @@ use Drupal\user\EntityOwnerTrait;
  *   admin_permission = "access instrument type overview",
  *   entity_keys = {
  *     "id" = "id",
- *     "label" = "label",
+ *     "label" = "code",
  *     "uuid" = "uuid",
  *     "owner" = "uid"
  *   },
@@ -51,6 +51,7 @@ class InstrumentType extends ContentEntityBase implements InstrumentTypeInterfac
 
   use EntityChangedTrait;
   use EntityOwnerTrait;
+  use ReclamationEntityFieldsHelper;
 
   /**
    * {@inheritdoc}
@@ -63,90 +64,12 @@ class InstrumentType extends ContentEntityBase implements InstrumentTypeInterfac
     }
   }
 
-  /**
-   * {@inheritdoc}
-   */
-  public static function baseFieldDefinitions(EntityTypeInterface $entity_type) {
+  protected static function fieldsMetadata() {
+    $fields = [];
 
-    $fields = parent::baseFieldDefinitions($entity_type);
-
-    $fields['label'] = BaseFieldDefinition::create('string')
-      ->setLabel(t('Label'))
-      ->setRequired(TRUE)
-      ->setSetting('max_length', 255)
-      ->setDisplayOptions('form', [
-        'type' => 'string_textfield',
-        'weight' => -5,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('view', [
-        'label' => 'hidden',
-        'type' => 'string',
-        'weight' => -5,
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['status'] = BaseFieldDefinition::create('boolean')
-      ->setLabel(t('Status'))
-      ->setDefaultValue(TRUE)
-      ->setSetting('on_label', 'Enabled')
-      ->setDisplayOptions('form', [
-        'type' => 'boolean_checkbox',
-        'settings' => [
-          'display_label' => FALSE,
-        ],
-        'weight' => 0,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('view', [
-        'type' => 'boolean',
-        'label' => 'above',
-        'weight' => 0,
-        'settings' => [
-          'format' => 'enabled-disabled',
-        ],
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['uid'] = BaseFieldDefinition::create('entity_reference')
-      ->setLabel(t('Author'))
-      ->setSetting('target_type', 'user')
-      ->setDefaultValueCallback(static::class . '::getDefaultEntityOwner')
-      ->setDisplayOptions('form', [
-        'type' => 'entity_reference_autocomplete',
-        'settings' => [
-          'match_operator' => 'CONTAINS',
-          'size' => 60,
-          'placeholder' => '',
-        ],
-        'weight' => 15,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('view', [
-        'label' => 'above',
-        'type' => 'author',
-        'weight' => 15,
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['created'] = BaseFieldDefinition::create('created')
-      ->setLabel(t('Authored on'))
-      ->setDescription(t('The time that the instrument type was created.'))
-      ->setDisplayOptions('view', [
-        'label' => 'above',
-        'type' => 'timestamp',
-        'weight' => 20,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('form', [
-        'type' => 'datetime_timestamp',
-        'weight' => 20,
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['changed'] = BaseFieldDefinition::create('changed')
-      ->setLabel(t('Changed'))
-      ->setDescription(t('The time that the instrument type was last edited.'));
+    $fields['code'] = self::simpleTextField(t('Code'));
+    $fields['liquidity'] = self::simpleTextField(t('Liquidity'));
+    $fields['instrument_type'] = self::simpleTextField(t('Instrument type'));
 
     return $fields;
   }
